requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
email-validator==2.1.0.post1
gevent-websocket==0.10.1
requests==2.31.0
geopy==2.3.0
shapely==2.0.1
pyproj==3.5.0
